--- api/lib/domain/usecases/change-user-lang.js
@@ -1,6 +1,0 @@
-const changeUserLang = async function ({ userId, lang, userRepository }) {
-  await userRepository.update({ id: userId, lang });
-  return userRepository.getFullById(userId);
-};
-
-export { changeUserLang };
